Generalize the complex mandatory keywords defintion to be used anywhere mandatory keywords are supported. Add support for mandatory_bio_keywords in set_skip_users
