+++ tests/util_tests.py
@@ -2,7 +2,11 @@ import unittest
 import logging
 from instapy.util import evaluate_mandatory_words
 
-class TestUtil(unittest.TestCase):
+# Note: This file is using a different name scheme than existing tests
+# so that unittest can be run with -p *_tests.py to discover all compatible tests in the tests directory
+
+
+class UtilsTests(unittest.TestCase):
     def setUp(self):
         pass
